requirements.txt
paho-mqtt==1.6.1
pandas==2.2.2
python-dateutil==2.8.2
PyYAML==6.0
